# Cache workload methods in simulator.