This PR implements a small optimization to cache workload methods rather than collecting the set of methods on every simulation. This doesn't matter much much compared to the other, bigger overheads. But, I'm optimizing this to get rid of some of the distracting overheads in profiles to help focus on the bigger problems.

```
$ go test -run=$^ -bench=. -count=5 | tee /tmp/baseline.txt
$ go test -run=$^ -bench=. -count=5 | tee /tmp/experiment.txt
$ benchstat /tmp/{baseline,experiment}.txt
name                        old time/op    new time/op    delta
Workloads/NoCallsNoGen-128    6.37ms ± 1%    6.34ms ± 2%    ~     (p=0.690 n=5+5)
Workloads/NoCalls-128         7.72ms ± 1%    7.84ms ± 2%    ~     (p=0.095 n=5+5)
Workloads/OneCall-128         16.7ms ± 3%    16.8ms ± 2%    ~     (p=0.841 n=5+5)

name                        old alloc/op   new alloc/op   delta
Workloads/NoCallsNoGen-128     513kB ± 0%     513kB ± 0%  -0.08%  (p=0.008 n=5+5)
Workloads/NoCalls-128          622kB ± 0%     621kB ± 0%  -0.08%  (p=0.008 n=5+5)
Workloads/OneCall-128         2.23MB ± 0%    2.22MB ± 0%    ~     (p=0.222 n=5+5)

name                        old allocs/op  new allocs/op  delta
Workloads/NoCallsNoGen-128     10.1k ± 0%     10.1k ± 0%  -0.12%  (p=0.008 n=5+5)
Workloads/NoCalls-128          14.1k ± 0%     14.1k ± 0%  -0.09%  (p=0.008 n=5+5)
Workloads/OneCall-128          45.4k ± 0%     45.4k ± 0%    ~     (p=0.222 n=5+5)
```